initial commit
Create an initial version of this action based on [1]. This adaptation
of that action will output the annotation of a git tag instead of the
message of a git message.

This commit contains a quick and dirty adaptation. Only changing the key
functionality in index.js, and changing the metadata files (README.md,
package.json, package-lock.json, and action.yml) as needed.

--
1. https://github.com/kceb/git-message-action