Add known issues, fix outputs, and improve writing
First, add a new section on known issues to the README. Particularly,
this describes the known issue with actions/checkout-at-v2 which does
not  preserve tag annotations correctly.

Second, this fixes the name of the output value of this Action in the
action.yml manifest file.

Lastly, some overall improvements to the wording of the description of
the Action.

And, as a bonus, fix the mailto: subject in SECURITY.md