Set up testing
Set up testing for the Action using Jest [1] and some simple manual
mocks for @actions/core and child_process.

The project has been slightly restructures to allow for testing. The
existing index.js file was nearly impossible to test because there is no
function to invoke. Hence, the logic of index.js has been moved to a
file in the (new) src/ directory. The only thing index.js does is call a
function that does everything it used to do.

The src/ directory was named by convention, and by the same convention
alll test files can be found in the test/ directory. Moreover, all test
files have the suffix ".test.js".

Extras:
- Fix a comment in .gitignore
- Fix mail address in SECURITY.md's mailto link

--
1. https://jestjs.io