Fix shell injection bug
Fix a shell injection bug where an attacker controlling the "tag" value
in the code (i.e. the value of `process.env.GITHUB_REF` or of the
`with: tag` input value) can run arbitrary (*) shell commands.

Also, update the Action's branding, changing the Action icon from
"git-commit" to "tag".

(*): provided it is within the limits of the GitHub Actions environment