Set up mutation testing
Set up mutation testing with Stryker [1]. Adding a configuration file,
npm command for running it as well as appropriate .gitignores.

Additionally, the test suites are updated to kill any mutants found by
Stryker. In particular, the "tries to get a tag from the input" test was
updated to verify that the correct input value is retrieved. Relatedly,
main.js was updated to get the input named "tag" only once to avoid
problems due to it being retreived twice.

Additionally, Jest is now configured to output its coverage report into
a new folder called "reports/", which is also where Stryker puts its
report.

Lastly, a clean command was added to return the repository to a clean
state.

--
1. https://stryker-mutator.io/