Don't use actions/checkout@v1 (#66)
* ci: run e2e test with checkout at v2

* ci: use `fetch-depth: 0` with checkout at v2

* docs: recommend using checkout at v2

Update the documentation to recommend using checkout at v2 instead of at
v1. The use of v1 is still suggested as a potential solution in the
scenario this Action isn't working, but it is (hopefully) clearly
communicated that this may be dangerous. Additionally, using v1 is
provided as a last resort alternative, fetching tags manually is
intentionally suggested first.