Set up code scanning/SAST with CodeQL
Add code scanning/SAST with CodeQL [1] through a new GitHub Actions [2]
workflow. This workflow is intended solely for code scanning/SAST jobs,
but currently only runs CodeQL.

The workflow is configured to only run when source code changes or when
scanner configurations change.

The permissions of the workflow are restricted to the minimum needed. As
an open source project, top-level `read-all` sets a sensible default.
The CodeQL job requires write permissions for security-events to be able
to report alerts to GitHub.

--
1. https://codeql.github.com/
2. https://github.com/features/actions