Report test coverage to Codecov
Start using Codecov [1] to report on coverage numbers. To achieve this
the verify workflow now has a separate unit test and coverage job. The
unit test job runs the unit test on each OS, the coverage job runs on
one OS and is intended to report the coverage the Codecov. A minimal
Codecov configuration is added the configure how it measures coverage,
what PR comments should look like, and what files to ignore.

--
1. https://codecov.io/