Update Continuous Integration (#74)
* Set up code scanning/SAST with CodeQL

Add code scanning/SAST with CodeQL [1] through a new GitHub Actions [2]
workflow. This workflow is intended solely for code scanning/SAST jobs,
but currently only runs CodeQL.

The workflow is configured to only run when source code changes or when
scanner configurations change.

The permissions of the workflow are restricted to the minimum needed. As
an open source project, top-level `read-all` sets a sensible default.
The CodeQL job requires write permissions for security-events to be able
to report alerts to GitHub.

* Use .nvmrc for NodeJS version in CI

To avoid having to update all instances manually.

There is one exception, which is the `test-e2` job, which should use the
same NodeJS version as is specified in `./action.yml`.

* Pin version of Actions used in workflows

Pin all used Actions in workflows to their current latest.

* Restrict CI runs

Restrict when the CI runs by only running CI workflows when relevant
files change.

`code-analysis.yml`: Was already restricted, but now only runs when
runtime code changes.

`update-index.yml`: Already as restricted as it can be.

`verify.yml`: Now only runs when code, code configurations, or
dependencies change.

* Setup a Stryker Dashboard

Set up the Stryker Dashboard [3] to have the mutation report uploaded
and easily accessible to anyone.

* Setup a Codecov Dashboard

Start using Codecov [4] to report on coverage numbers. To achieve this
the verify workflow now has a separate unit test and coverage job. The
unit test job runs the unit test on each OS, the coverage job runs on
one OS and is intended to report the coverage the Codecov. A minimal
Codecov configuration is added the configure how it measures coverage,
what PR comments should look like, and what files to ignore.

* Add continuous integration badges to README

--
1. https://codeql.github.com/
2. https://github.com/features/actions
3. https://stryker-mutator.io/docs/General/dashboard/
4. https://codecov.io/