Check callCount rather than calledOnce for better errors
When using

    const stub = sinon.stub();
    assert.ok(stub.calledOnce);

the test error will simply say "false is not true" without an indication
of whether the stub was called at all or not. Checking the `callCount`
instead avoids this by comparing the number of calls, e.g.

    const stub = sinon.stub();
    assert.is(stub.callCount, 1);

will tell you whether the stub was called 0, 2, or some other number of
times.