Fix git hooks when there's no stashes (#83)
The previous implementation would fail if there's no stashes because the
command exits with an error in that scenario. This is fixed by
redirecting stderr to `/dev/null` (to prevent the error from showing up
in the hook output) and explicitly returning `"0"` if the output to
stdout is nothing. As this is more complex than a single command, the
logic is extracted to a separate file and used in every place that
previously just used the stash-count command.