Fix test:mutation command
Update the commandRunner to pass equivalent arguments to AVA as were
passed to Jest. The additional argument prevents AVA from doing to much
parallelism which may reduce performance when combined with Strykers
parallelism.

Additionally, somehow Stryker doesn't work when combined with AVA and is
missing the `at-babel/travers` package. To circumvent this issue it's
manually added as a devDependency here.