Fix test:mutation command
Update the commandRunner to omit all custom arguments. In constract to
Jest, uvu doesn't offer a way to reduce the amount of concurrency.
Luckily, the test suite for this project is minimal enough that the
concurrency of the test suite has minimal interference with the
concurrency of Stryker.

Additionally, somehow Stryker doesn't work when combined with uvu and is
missing the `at-babel/travers` package. To circumvent this issue it's
manually added as a devDependency here.