Update coverage command
Update the coverage command to leverage c8 [1] to generate a coverage
report after running the tests. Two reporters are enabled for c8, one to
output an lcov report (.info and HTML) and one to output a text-based
report on the terminal.

One alternative was considered, namely nyc [2]. c8 was chosen over nyc
because they offer similar functionality but c8 is more modern and
up-to-date.

--
1. https://github.com/bcoe/c8
2. https://istanbul.js.org/