Improve git hooks (#88)
* Don't create a stash during merging

Update the git hooks to *not* create a stash during merging, this is
because stashing during merging messes up the merge process. So, now it
is alwasy necessary to use the `--no-verify` flag to disable the hooks
when committing a merge.

* Update git hook formatting

Prefer the use of

    if [ ... ]; then

over

    if [ ... ]
    then

because it's fewer lines and follows the if-formatting used in the
JavaScript code of the project (`if (...) {` not `if (...)\n{`).

* Stash unstaged *and* staged changes before pushing

In contrast to when you're committing, when you're pushing we want to
know if the current state of the project is "correct" (according to the
checks in the hook). (When committing we want to know if the next state,
including the staged changes, is "correct").