Test with uvu instead of Jest (#82)
* Refactor main to avoid having to mock modules

* Rewrite tests using uvu

Remove Jest (dependency and mocks) and rewrite the test suite using
uvu [1] and Sinon.JS [2].

* Check callCount rather than calledOnce for better errors

When using

    const stub = sinon.stub();
    assert.ok(stub.calledOnce);

the test error will simply say "false is not true" without an indication
of whether the stub was called at all or not. Checking the `callCount`
instead avoids this by comparing the number of calls, e.g.

    const stub = sinon.stub();
    assert.is(stub.callCount, 1);

will tell you whether the stub was called 0, 2, or some other number of
times.

* Update coverage command

Update the coverage command to leverage c8 [3] to generate a coverage
report after running the tests. Two reporters are enabled for c8, one to
output an lcov report (.info and HTML) and one to output a text-based
report on the terminal.

One alternative was considered, namely nyc [4]. c8 was chosen over nyc
because they offer similar functionality but c8 is more modern and
up-to-date.

* Update test:mutation command

Update the `commandRunner` to omit all custom arguments. In contrast
to Jest, uvu doesn't offer a way to reduce the amount of concurrency.
Luckily, the test suite for this project is minimal enough that the
concurrency of the test suite has minimal interference with the
concurrency of Stryker.

--
1. https://github.com/lukeed/uvu
2. https://sinonjs.org
3. https://github.com/bcoe/c8
4. https://istanbul.js.org/