Add script to update the CHANGELOG
Add a simple script that can be invoked to update the changelog. The
script alters the changelog *only* if 1) the manifest specifies a
version not found in the CHANGELOG, and 2) there are unreleased changes.
If thsoe conditions hold, it sets the unreleased changes as changes
released in the version found (at time of running) in the package
manifest.

Update the Release Guidelines to recommend using this script.