Add CI job to commit a version bump and open a Pull Request for it
Add a new job and "Release" workflow to automate the version bump
commit. To prevent having to push to the `main` branch (allowing for
full branch protections) and to encourage review before accepting an
automated change, the version bump is committed to a separate branch and
a Pull Request is opened (both) through peter-evans/create-pull-request
[1].

This essentially automates step 1, 2, 3, and ~5 of the Release
Guidelines through means of a Pull Request.

--
1. https://github.com/peter-evans/create-pull-request