Detect if files in lib/ are staged, inform about restoring
Add a basic check to the pre-commit hook that checks if there are any
changes to the files in the lib/ directory and informs the committer 1)
1) that those will be unstaged before committing, and 2) why that is the
case.

The LIB_STAGED_COUNT value is computed by getting the names (`--name
-only`) from the diff (`git diff`) of staged (`--staged`) files and
counting the number of lines. If there are no staged files in lib/ there
will be 0 lines. If there are any staged files in lib/ there will be
more than 0 lines.