Automate the release process (#98)
* Add script to update the CHANGELOG

Add a simple script that can be invoked to update the changelog. The
script alters the changelog *only* if 1) the manifest specifies a
version not found in the CHANGELOG, and 2) there are unreleased changes.
If thsoe conditions hold, it sets the unreleased changes as changes
released in the version found (at time of running) in the package
manifest.

Update the Release Guidelines to recommend using this script.

* Update guidelines to bump version

Update the "Updating the Version Number" section of the Release
Guidelines to include a command that can be used to apply a version bump
in the manifest and lockfile.

* Add CI job to commit a version bump and open a Pull Request for it

Add a new job and "Release" workflow to automate the version bump
commit. To prevent having to push to the `main` branch (allowing for
full branch protections) and to encourage review before accepting an
automated change, the version bump is committed to a separate branch and
a Pull Request is opened (both) through peter-evans/create-pull-request
[1].

This essentially automates step 1, 2, 3, and ~5 of the Release
Guidelines through means of a Pull Request.

--
1. https://github.com/peter-evans/create-pull-request

* Update lib/ in release PR instead of on `main`

Remove the workflow that keeps the build files (lib/) up-to-date on the
`main` branch and replace it with including the build output in the
release Pull Request as part of the version bump commit. This will
result in fewer commits and even lower chance of merge conflicts.

* Update the Release Guidelines

* Automate tag creation for new versions

Automate steps 3-7 of the "automated release" process by automatically:

1. Detecting if a new version needs to be released by seeing if a tag
   exists for the version specified in the package manifest, and if so
2. Creating the tag for the new version.
3. Pushing the tag for the new version to GitHub.
4. Updating the v1 tag to point to the latest commit.
5. PUshing the v1 tag to GitHub.

for all pushes on the `main` branch. The new CI job will not make any
changes if there is already a tag for the version in the package
manifest.

The Release Guidelines are updated accordingly.

* Update Release Guidelines

Clarify that the manual release process should also be referenced when
something goes wrong with the automatic release process.