Better focus CI jobs (#100)
* Only run verify code when source code or tests change

... not when build files (`lib/`) or scripts (`script/`) change.

* Run linting CI when JSON files change

Matching the prettier configuration. Also update the lint/format npm
scripts to prevent duplication.

* Run Verify code workflow when Action manifest changes

Run the verify workflow whenever the Action manifest (`action.yml`)
changes so that the e2e tests run and gain confidence that the changes
don't break the Action.

* Run Verify code workflow when the Codecov configuration changes

Run the verify workflow whenever the Codecov configuration
(`.github/codecov.yml`) changes so that the tests with coverage run and
report changes to Codecov to be able to check that the changes take
affect and make sense.