Add step-security/harden-runner to harden GH Actions (#148)
* Add step-security/harden-runner to workflows

Add the step-security/harden-runner [1] Action to this project's GitHub
Actions Workflows (except the release workflow for now, because it's not
run all the time and therefor not useful for testing purposes). The
latest version at the time of committing is pinned.

The purpose of this Action is to first detect what _should_ happen
during a typical job run, and then enforce that with specialized config
after that. E.g., detect what HTTP traffic should happen, and then block
all other HTTP traffic in subsequent runs based on the first run.

* Don't try to run step-security/harden-runner on MacOS or Windows

This is currently unsupported [2].

* Configure step-security/harden-runner

Set egress-policy to block anything not in the allowlist and configure
the allowlist based on egress detected by the Action during audit. Also
add a name to the step-security/harden-runner job step.

* Add step-security/harden-runner in audit mode to release workflow

So that it can audit the workflow when it's run (with the next release)
and can be configured afterwards.

--
1. https://github.com/step-security/harden-runner
2. https://github.com/step-security/harden-runner/blob/56008243b5bcc21d227601d398064599c69eb33c/src/index.ts#L6