Add scheduled workflow to audit dependencies (#199)
* Add scheduled workflow to audit dependencies

Add a new GitHub Actions workflow for auditing dependencies. It is
scheduled weekly for an initial trail, and audits dependencies on both
`main` and `main-v1`. The latter is especially of interest, as it a
failed job would alert about problems on v1 that are not present on v2
(since `npm audit` exits with a non-zero exit code if a problem is
detected). The former is more-or-less already covered by Dependabot.

* Configure harder-runner for dependency analysis

Based on [1] and prior experience.

--
1. https://app.stepsecurity.io/github/ericcornelissen/git-tag-annotation-action/actions/runs/2364201254