Update Node.js version for dev to ^18 and npm to ^8
Update the Node.js version to be used for development to 18 (or higher)
by updating `.nvmrc` and `package.json`. Update the npm version to be
used during development to 8 (or higher) by updating `package.json`. The
Contributing Guidelines are updated accordingly. (Also updated the
formatting of "Node.js" and "npm" in the Contributing guidelines in line
with the official formatting, based on the website `<title>`s.)