Update Node.js version to ^18 and npm to ^8 for development (#211)
* Update Node.js version for dev to ^18 and npm to ^8

Update the Node.js version to be used for development to 18 (or higher)
by updating `.nvmrc` and `package.json`. Update the npm version to be
used during development to 8 (or higher) by updating `package.json`. The
Contributing Guidelines are updated accordingly. (Also updated the
formatting of "Node.js" and "npm" in the Contributing guidelines in line
with the official formatting, based on the website `<title>`s.)


* Upgrade to lockfile version 3


* Update installed Node.js version for e2e tests

Change the Node.js version that is installed during the e2e test job
from a version matching `./action.yml` to simply the version present in
`./.nvmrc`. The thinking behind the pre-existing approach was that the
version should match `./action.yml` to more closely match a "real"
Action execution. However, during a "real" execution of the Action the
Node.js version being installed might also be different from the version
specified in the Action manifest, hence I don't think it actually makes
sense to require the installed version to match the Action runtime
version.

This is changed in particular here because the Node.js and npm version
used for development are being upgraded to a version not matching what's
present in `./action.yml`, which leads to the CI failure on the `npm ci`
step because of the `"engine"` requirements specified in `package.json`
(in combination with the `engine-strict=true` set in `./.npmrc`).


* Allow "nodejs.org:443" in CI

Update the step-security/harden-runner configuration to allow requests
to "nodejs.org" to be able to download Node.js binaries.