Add depcheck and unimported for vet command
Add a npm command called "vet" for vetting the source code of this
project. Vetting is currently powered by depcheck [1] and unimported [2]
which are both projects that analyse source code for problems related to
dependencies. In particular, the former looks for unused dependencies
declared in the package manifest. The latter looks for unused source
code files.

--
1. https://www.npmjs.com/package/depcheck
2. https://www.npmjs.com/package/unimported