Initialize source code vetting (#215)
* Add depcheck and unimported for vet command

Add a npm command called "vet" for vetting the source code of this
project. Vetting is currently powered by depcheck [1] and unimported [2]
which are both projects that analyse source code for problems related to
dependencies. In particular, the former looks for unused dependencies
declared in the package manifest. The latter looks for unused source
code files.


* Continuously vet source code in the CI


* Configure depcheck and unimported

Add configuration files for both depcheck [1] and unimported [2].

For depcheck, notable is line 2. Rollup is explicitly ignored here
because it's usage is not detected in package.json at line 11.

For unimported, the configuration is just tailored to the needs of this
project.


* Add vetting to git hooks

Run the vet command both for each commit (to ensure each commit is
proper) as well as each push (to double ensure, in case `--no-verify`
was used).


--
1. https://www.npmjs.com/package/depcheck
2. https://www.npmjs.com/package/unimported