Don't run gitleaks-action on schedule (#252)
Prevent git gitleaks-action job from running if the workflow is
triggered by the 'schedule' event. This is because the 'schedule' event
is not (currently) supported by gitleaks-action and errors
unexpectedly when run on schedule. This change may be reverted in the
future if the 'schedule' event is supported by gitleaks-action.