Update executable JS file included in VSC
Rebuild the executable JS file to update it from the `.js` file to the
(new) `.cjs` file.

The .gitignore allowlist has been moved from the root to the lib/ dir to
increase the proximity and hopefully improve understandability and
correctness of the allowlist.

Also partially reverting 923289e344f2771328d632b0fc8a0e6ff456d0d2 as
the config changes turned out to be incorrect as it lead to node_modules
depenencies not being bundled into the executable JavaScript file.