Prefer property tests over parameterized tests
Replace the pre-existing parameterized tests (tests that use a for-loop
to execute the same tests with multiple input values, aka parameters)
with property based tests (where the input values are randomly generated
rather than pre-defined). This should increase confidence in the
correctness of the program as it's not possible to hardcode the values
being tests in the program (removing one avenue for errors).