Update the Release Guidelines
The primary change is an update to the manual release steps to encourage
the use of a release branch + Pull Request over pushing to `main`
directly.

Additionally, separate running `npm run build` from the "verify release
state" step into its own step so that it can be clarified why it has to
be run.