Update e2e tests (#304)
* Use exact string match in e2e test

To guarantee the expected outcome without having to inspect the logs to
verify the expected string was outputted by the action, rather than just
a non-empty one.

* Set `fail-fast: false` for e2e tests

As it's not guaranteed the e2e tests will fail for each OS if it fails
for one, the failure may be OS dependent. Hence, it's useful to have the
e2e tests complete for all OSes.

* Remove useless echo

Originally this echo was included to allow for easily viewing the output
of the action. However, this can be seen in the logs anyway.