Use Fossa CLI to check licenses
Adjust the CI and project scripts with a defined approach for checking
dependency licenses using the Fossa CLI. This indicates a slight change
from the previous setup, where the Fossa GitHub integration was used.
Using the CLI makes it possible to support somewhat offline[^1] checking
of licenses.

Note that, for now, installing the Fossa CLI is a manual process that's
on the contributor. The Contributing Guidelines explain this, as well as
how to check the licenses locally.

--
[^1]: Somewhat because it still requires a Fossa account and an internet
connection.