Create devcontainer specification for this project
Define a devcontainer for the project based on a custom Dockerfile. The
Dockerfile is used primarily as a convenient way to install some extra
development dependencies. The devcontainer is entirely based on what the
Contributing Guidelines say one needs to contribute. Based on my testing
this is actually the case.

The .devcontainer directory is used because the Dockerfile isn't really
usable on it's own.

Why .devcontainer? Primarily because I wanted to give it a try :)