Store Stryker's incremental file in a cache dir
Since the incremental file is essentially a cache file for Stryker, and
isn't intended for humans to read, it doesn't make much sense to put it
in the _reports/ dir.

Instead, put it in a new .cache/ dir. The git config as well as clean
script were updated accordingly.

The .temp/ dir was also considered for this, however I feel that's
semantically incorrect.