General minor changes (#364)
* Correct minimum npm version in Contributing Guidelines

The minimum version is 8.1.2 because prior to that version (on v8) the
option `lockfile-version=3` from .npmrc doesn't work.

* Instruct Stryker to remove temporary dir if successful

Per the Stryker documentation:

> Choose whether or not to clean the temp dir (which is ".stryker-tmp"
> inside the current working directory by default) after a successful
> run. The temp dir will never be removed when the run failed for some
> reason (for debugging purposes).

So, since on failure it will remain for debugging, and debugging isn't
really necessary on success (the report provides all necessary info),
the temporary directory should be cleaned.

* Store Stryker's incremental file in a cache dir

Since the incremental file is essentially a cache file for Stryker, and
isn't intended for humans to read, it doesn't make much sense to put it
in the _reports/ dir.

Instead, put it in a new .cache/ dir. The git config as well as clean
script were updated accordingly.

The .temp/ dir was also considered for this, however I feel that's
semantically incorrect.

* Update Stryker incremental report location in CI cache