Replace Fossa by licensee for license checking
Remove the dependency on Fossa (config, CI, and Contributing Guidelines)
and update the `check-license` script to use licensee instead. The CI
and Contributing Guidelines have been updated accordingly. An initial
configuration for licensee is added, but it's incomplete for now in
order to demonstrate a failure of the check in the CI.