Use `licensee` for license scanning (#365)
* Add licensee as a dependency

Add the licensee package [1] as a dependency to be used for validating
the licenses of npm dependencies.

* Replace Fossa by licensee for license checking

Remove the dependency on Fossa (config, CI, and Contributing Guidelines)
and update the `check-license` script to use licensee instead. The CI
and Contributing Guidelines have been updated accordingly. An initial
configuration for licensee is added, but it's incomplete for now in
order to demonstrate a failure of the check in the CI.

* Configure licensee

Update the licensee [1] configuration for the current state of the
project. First, the list of approved licenses has been extended to
accommodate (almost) all dependencies of this project.

Second, there's two "exceptions" in the `"packages"` section. These are
here because:
- `docopt`: Doesn't specify its license correctly (i.e. in accordance
  with the spec of npm `package.json`). Manual review shows it uses the
  MIT license:

      cat node_modules/docopt/package.json | grep "license"

- `scss-parser`: Doesn't specify the license in the `package.json`,
  instead referring to the README. Manual review of the license shows
  it's licensed under the "BSD-3-Clause" license (see `## License`
  section):

      cat node_modules/scss-parser/README.md

* Configure harder-runner for new license check 

--
[1]: https://www.npmjs.com/package/licensee