Setup hadolint-based Dockerfile linting
Add a script to run hadolint [1] on the devcontainer Dockerfile for this
project. Document this command in the Contributing Guidelines (note that
no new prerequisite is needed as hadolint is being invoked through
Docker). Finally, add an initial configuration file for hadolint,
basically just to be explicit about some options.

--
1. https://github.com/hadolint/hadolint