Automatically create a GitHub release for new releases (#385)
* Automatically create GitHub releases for new releases

Immediately after pushing the tag, create a GitHub release for it. The
release's title will be in accordance with the Release Guidelines. The
release text is currently still missing.

* Create and use script to get release notes

Create script to get release notes for the current version and use it in
the release workflow to set the GitHub Release body.