Pin all runner images in GitHub Actions workflows (#420)
* Pin all runner images in GitHub Actions workflows

Update all GitHub Actions workflows to use a specific runner image for
all jobs. The chosen images are based on what the documented `latest`
images are on December 9, 2022. Upgrading to more recent images may be
done seperately.

Doing this should help with reproducability. It's currently not known if
Dependabot will keep these up-to-date automatically, but since upgrades
for these are generally uncommon manually keeping them up-to-date should
be doable.

* Remove conditional harden-runner for os-matrix job

Based on workflows in other projects, harden-runner won't fail for
operating systems it doesn't support. So, the condition can be removed
safely. This also means harden-runner will automatically become active
on other OSes if support is added.

* Make OS matrixed job names constant

Update matrix jobs that vary the OS to have a constant name while the os
version can change over time.

The primary advantage of doing this is that the job can be made required
and required jobs don't need to change if the OS version is changed.