Update `lint:docker` command (#438)
Change the `lint:docker` command details to run from a version tag
instead of an image hash. Image hashes are tied to the target
architecture, making it inconvenient for this particular use case.
Switching to a version tag still provides good reproducability
guarantees (it can be expected that a version isn't re-published) - in
contrast to an image hash which provides better reproducability
guarentees (it makes re-publishing cryptographically difficult) and by
extension more secure.