Create nightly workflow to update tooling
Create a GitHub Actions workflow to update tooling nightly. This
workflow will use asdf to install the latest version of each tool,
update the `.tool-versions` file if necessary, and create a Pull
Request to update the tool (if there's any changes).

For convenience, the workflow can be dispatched manually as well. This
can come in handy if an update is available and it is needed
immediately.

Note that the list of tools is currently duplicated between the workflow
and the `.tool-versions` file. This may be improved upon in the future.