Use `nick-fields/retry` to avoid tooling update flakiness
Update the nightly workflow's tooling update job to automatically retry
the flaky step of getting the latest available version of the tool using
`nick-fields/retry` [1]. Since the duration of this job doesn't matter
much, and to be on the safe side, a timeout of 1 minute is configured to
increase the probability of success.

Since the outputs of `nick-fields/retry` are restricted, the latest
version is now stored as an environment variable.

--
1. https://github.com/nick-fields/retry