Reconfigure `nick-fields/retry`
Per one of the issues on the projects repository, the last command must
fail in order for the command to be retried (fix for this is pending).

Additionally, update the configuration to set `retry_wait_seconds`.
`timeout_minutes` was previously set thinking it determined the time
between retries, but that's actually what `retry_wait_seconds`.
`timeout_minutes` is kept to retry if the network hangs for some reason
(correspondingly, `retry_on: error` has been removed to allow for
retries in such scenarios).

Also make the command invoked incorrect for testing purposes. This
should be reverted.