Adjust retry timings for getting latest version of tools
Increase retry wait time. 30 seconds proved to not be sufficient for
getting success on the second try, trying 2 minutes now. As this is run
nightly it doesn't matter too much if this takes 4 minutes and a bit to
succeed.

Also reduce the timeout to 20 seconds, in general, if successful, this
steps takes only a couple of seconds.