Update cosign usage
Update the usage of cosign following the upgrade from v1 to v2. Per the
changelog [1], "COSIGN_EXPERIMENTAL=1 is no longer required to have
identity-based ("keyless") signing and transparency" and "Verification
now requires identity flags" (the change "verify-blob no longer searches
for a certificate" does not apply as `--certificate` is already used).

I tested these changes, but am not sure this `--certificate-identity`
will be valid if the signature is created by the workflow. As such, it
may be necessary to manually create the next GitHub release _if_ the
verification step fails.

--
1. https://github.com/sigstore/cosign/blob/69c9b37/CHANGELOG.md