Set up continuous SAST with Semgrep (#502)
* Set up continuous SAST with Semgrep

Add a new Static Application Security Test (SAST) scan with Semgrep [1].
Note that Semgrep supports scanning Shell scripts (most relevant to this
project) as well as configuration files such as GitHub Actions workflows

Harden-runner is not configured for this job because it doesn't work
with container-based jobs.

* Harden `release.yml` workflow

Prevent shell injection in this workflow by capturing the update type as
an environment variable and using the environment variable. This way, the
expansion of the input in the command can't result in shell injection.

Yes, this particular case is not very vulnerable due to 1) limited value
space, and 2) trusted triggers. However, it's hard to guarantee those
variables stay fixed and simple to apply the fix.

This problem was detected by Semgrep [1] using a full-repository scan.

--
1. https://semgrep.dev