Generate and upload Semgrep SARIF report (#513)
Update the "Check / Semgrep" job to have Semgrep generate an output in
SARIF format and upload that to GitHub using
`github/codeql-action/upload-sarif` (which requires the added
permissions).

Also don't run Semgrep for PRs by Dependabot as it doesn't have access
to `secrets.SEMGREP_APP_TOKEN`. In turn the `semgrep` invocation doesn't
do anything, as a result of which uploading the SARIF file will fail
(leading to an job failure).