Drop Semgrep dependency (#554)
Remove the CI job that runs Semgrep from the project. The primary reason
for this is to reduce the number of project dependencies. While Semgrep
works pretty well and isn't noisy, it doesn't add too much to this
project at the moment.

As an added bonus, this also removes all dependency on the CodeQL
Action project.