Remove release notes automation
Drop the Node.js prerequisite by removing the automation that pulls the
changes from the changelog and puts it in the release notes. This is
simpler than trying to rewrite it without JavaScript and is currently
considered acceptable because of limited impact due to the infrequency
of changes.

A contribution in the form of a script to replace the removed
`get-release-notes.mjs` script that does not add a prerequisite are
welcome.